+++ .github/workflows/main.yml
@@ -4,15 +4,15 @@ on:
   push:
   pull_request:
   schedule:
-    - cron:  '0 0 * * *'
+    - cron: "0 0 * * *"
 
 jobs:
   validate:
     runs-on: "ubuntu-latest"
     steps:
-        - uses: "actions/checkout@v3"
-        - uses: "home-assistant/actions/hassfest@master"
-        
+      - uses: "actions/checkout@v3"
+      - uses: "home-assistant/actions/hassfest@master"
+
   hacs:
     name: HACS Validation
     runs-on: ubuntu-latest
