+++ custom_components/ha_gas_station_spain/const.py
@@ -1,0 +1,12 @@
+"""Constants for Gas Station Spain."""
+
+DOMAIN = "gas_station_spain"
+
+UPDATE_INTERVAL = 2
+CONF_PROVINCE = "province"
+CONF_PRODUCT = "product"
+CONF_MUNICIPALITY = "municipality"
+CONF_STATION = "station"
+CONF_FIXED_DISCOUNT = "fixed_discount"
+CONF_PERCENTAGE_DISCOUNT = "percentage_discount"
+CONF_SHOW_IN_MAP = "show_in_map"
