+++ custom_components/ha_gas_station_spain/manifest.json
@@ -6,6 +6,6 @@
   "documentation": "https://github.com/MiguelAngelLV/gas_station_spain",
   "iot_class": "cloud_polling",
   "issue_tracker": "https://github.com/MiguelAngelLV/gas_station_spain/issues",
-  "requirements": [],
-  "version": "0.8.0"
+  "requirements": ["gas-station-spain==1.1.0"],
+  "version": "2.0.0"
 }
