+++ custom_components/ha_gas_station_spain/translations/pt.json
